refactor User putting newsletter in concerns
